Promote daObjWc_Obj03_c to a single-file TU
Consolidates the ten one-symbol `src/_ZN15daObjWc_Obj03_c*.cpp` shards and
`src/daObjWc_Obj03_c_classInit.c` into `src/actors/d_a_obj_wc_obj03.cpp`,
written in reverse ROM order.

The previous shadow in `src_tu/` carried its own "THIS TU MUST NOT BE
PROMOTED" banner -- ordinals [2], [3] and [4] had been degraded to
`func_<addr>` raw-pointer-arithmetic placeholders -- so the TU was
reconstructed from the real named-member sources instead.

Two conditions, the same pair the sibling daObjWc_Mizu_c settled:

* Emission order. The cartridge keeps D1 at 0x02111728 below D0 at
  0x02111760, while a real destructor member makes mwccarm emit the
  D2/D1/D0 triple as one D0-then-D1 group. Both variants are carried as
  `// @symbol` mangled bodies; the two CONVERTED criteria that costs are
  banked in `converted-backslide-exceptions.jsonl`.
* No vptr bias. `_ZTV15daObjWc_Obj03_c` at 0x02113db4 is the address
  point -- D1's literal pool word at 0x0211175c is exactly that -- so
  every vptr store names the symbol with addend 0.

`include/daObjWc_Obj03_c.h` moves the destructor declaration after the
other members so mwccarm still emits the class's vtable and RTTI as vague
linkage with no key function to home them; the nine resulting data
symbols (this class's three plus the dActor_c/fBase_c/dBase_c passengers)
are licensed as `compiler_only_output`. Declared first, the TU emits none
of them and the romdata ratchet reds the PR.

Also drops a duplicate `_ZTV15daObjWc_Obj03_c` extern left in
`decl_common.h` by the class rename.

Proof: `tools/rombuild.py -j16` -- 106/106 exact, 100.000000%,
mismatching 0, ROM data from source 606 verified.
