+++ attribution.json
@@ -1327,6 +1327,16 @@
     "src/actors/d_a_obj_wc_mizu.cpp#_ZN14daObjWc_Mizu_c6RenderEv": "github-actions[bot]",
     "src/actors/d_a_obj_wc_mizu.cpp#_ZN14daObjWc_Mizu_c8BehaviorEv": "github-actions[bot]",
     "src/actors/d_a_obj_wc_mizu.cpp#_ZN14daObjWc_Mizu_c13InitResourcesEv": "github-actions[bot]",
-    "src/actors/d_a_obj_wc_mizu.cpp#daObjWc_Mizu_c_classInit": "github-actions[bot]"
+    "src/actors/d_a_obj_wc_mizu.cpp#daObjWc_Mizu_c_classInit": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_cD1Ev": "andrewboudreau",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_cD0Ev": "andrewboudreau",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c19CheckClsnWithPlayerEv": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c10SetWaterIDEv": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c20UpdateModelTransformEv": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c16CleanupResourcesEv": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c6RenderEv": "andrewboudreau",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c8BehaviorEv": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c13InitResourcesEv": "github-actions[bot]",
+    "src/actors/d_a_obj_wc_obj03.cpp#daObjWc_Obj03_c_classInit": "github-actions[bot]"
   }
 }
