+++ config/arm9/overlays/ov029/delinks.txt
@@ -12,45 +12,9 @@ src/actors/d_a_obj_wc_obj02.cpp:
     complete
     .text start:0x0211137c end:0x02111728
 
-src/_ZN15daObjWc_Obj03_cD1Ev.cpp:
+src/actors/d_a_obj_wc_obj03.cpp:
     complete
-    .text start:0x02111728 end:0x02111760
-
-src/_ZN15daObjWc_Obj03_cD0Ev.cpp:
-    complete
-    .text start:0x02111760 end:0x021117ac
-
-src/_ZN15daObjWc_Obj03_c19CheckClsnWithPlayerEv.cpp:
-    complete
-    .text start:0x021117ac end:0x02111850
-
-src/_ZN15daObjWc_Obj03_c10SetWaterIDEv.cpp:
-    complete
-    .text start:0x02111850 end:0x021118c8
-
-src/_ZN15daObjWc_Obj03_c20UpdateModelTransformEv.cpp:
-    complete
-    .text start:0x021118c8 end:0x02111908
-
-src/_ZN15daObjWc_Obj03_c16CleanupResourcesEv.cpp:
-    complete
-    .text start:0x02111908 end:0x0211192c
-
-src/_ZN15daObjWc_Obj03_c6RenderEv.cpp:
-    complete
-    .text start:0x0211192c end:0x02111954
-
-src/_ZN15daObjWc_Obj03_c8BehaviorEv.cpp:
-    complete
-    .text start:0x02111954 end:0x02111a04
-
-src/_ZN15daObjWc_Obj03_c13InitResourcesEv.cpp:
-    complete
-    .text start:0x02111a04 end:0x02111a84
-
-src/daObjWc_Obj03_c_classInit.c:
-    complete
-    .text start:0x02111a84 end:0x02111ac4
+    .text start:0x02111728 end:0x02111ac4
 
 src/actors/d_a_obj_wc_obj05.cpp:
     complete
