+++ config/converted-backslide-exceptions.jsonl
@@ -172,3 +172,5 @@
 {"path": "src/actors/d_a_obj_wc_obj04.cpp#_ZN15daObjWc_Obj04_cD1Ev", "reason": "The cartridge keeps D1 at 0x02112630 and D0 at 0x0211267c -- D1 BELOW D0 -- and mwccarm emits one .text section per function in the reverse of source order, so this TU has to place the two bodies independently. Defining ~daObjWc_Obj04_c() once as a real C++ member instead was tried and refused by the build: the D2/D1/D0 triple emits as a group in the order D0-then-D1, which objisolate rejects with 'requested functions are not emitted in ROM order' (and adds an unhomed D2). So D1 and D0 stay as '// @symbol' marked bodies under their mangled names, which is what costs them the no-raw-offset and no-mangled-refs criteria. This is goal 1 (byte-match) over goal 3 (readability), the trade tiers_ratchet exists to make loud. The clean one-line member versions the promotion deletes are recoverable from git history if the placement rule ever changes."}
 {"path": "src/actors/d_a_obj_wc_mizu.cpp#_ZN14daObjWc_Mizu_cD0Ev", "reason": "ov029/daObjWc_Mizu_c promotion: the cartridge keeps D1 at 0x021121a4, BELOW D0 at 0x021121f0, while a real `daObjWc_Mizu_c::~daObjWc_Mizu_c()` member makes mwccarm emit the D2/D1/D0 triple as one group in the order D0-then-D1. objisolate then refuses the TU with `requested functions are not emitted in ROM order`, and the group also carries an unhomed D2. Writing D1 and D0 as separate `// @symbol` marked mangled bodies is the only placement that reaches both ROM addresses, and it costs this tier's no-raw-offset and no-mangled-refs criteria. Byte-match over readability; same wall as ov029/daObjWc_Obj04_c (#2183)."}
 {"path": "src/actors/d_a_obj_wc_mizu.cpp#_ZN14daObjWc_Mizu_cD1Ev", "reason": "ov029/daObjWc_Mizu_c promotion: the cartridge keeps D1 at 0x021121a4, BELOW D0 at 0x021121f0, while a real `daObjWc_Mizu_c::~daObjWc_Mizu_c()` member makes mwccarm emit the D2/D1/D0 triple as one group in the order D0-then-D1. objisolate then refuses the TU with `requested functions are not emitted in ROM order`, and the group also carries an unhomed D2. Writing D1 and D0 as separate `// @symbol` marked mangled bodies is the only placement that reaches both ROM addresses, and it costs this tier's no-raw-offset and no-mangled-refs criteria. Byte-match over readability; same wall as ov029/daObjWc_Obj04_c (#2183)."}
+{"path": "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_cD0Ev", "reason": "ov029/daObjWc_Obj03_c promoted to src/actors/d_a_obj_wc_obj03.cpp: D1/D0 carried as // @symbol mangled bodies for the ROM emission order (D1 below D0), costing the no-raw-offset and no-mangled-refs criteria for those two functions"}
+{"path": "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_cD1Ev", "reason": "ov029/daObjWc_Obj03_c promoted to src/actors/d_a_obj_wc_obj03.cpp: D1/D0 carried as // @symbol mangled bodies for the ROM emission order (D1 below D0), costing the no-raw-offset and no-mangled-refs criteria for those two functions"}
