+++ config/converted-baseline.json
@@ -7,7 +7,7 @@
     "no_codegen_trick",
     "no_mangled_refs"
   ],
-  "count": 2588,
+  "count": 2586,
   "converted": [
     "src/ARMMathLoadState.c",
     "src/ARMMathSaveState.c",
@@ -576,13 +576,6 @@
     "src/_ZN12PiranhaPlant6RenderEv.cpp",
     "src/_ZN12PiranhaPlantD0Ev.cpp",
     "src/_ZN12PiranhaPlantD1Ev.cpp",
-    "src/_ZN15daObjWc_Obj03_c10SetWaterIDEv.cpp",
-    "src/_ZN15daObjWc_Obj03_c16CleanupResourcesEv.cpp",
-    "src/_ZN15daObjWc_Obj03_c19CheckClsnWithPlayerEv.cpp",
-    "src/_ZN15daObjWc_Obj03_c20UpdateModelTransformEv.cpp",
-    "src/_ZN15daObjWc_Obj03_c6RenderEv.cpp",
-    "src/_ZN15daObjWc_Obj03_cD0Ev.cpp",
-    "src/_ZN15daObjWc_Obj03_cD1Ev.cpp",
     "src/_ZN12WaterSuction16CleanupResourcesEv.cpp",
     "src/_ZN12WaterSuction16OnPendingDestroyEv.cpp",
     "src/_ZN12WaterSuction6RenderEv.cpp",
@@ -2433,6 +2426,11 @@
     "src/actors/d_a_obj_wc_obj02.cpp#_ZN15daObjWc_Obj02_c6RenderEv",
     "src/actors/d_a_obj_wc_obj02.cpp#_ZN15daObjWc_Obj02_cD0Ev",
     "src/actors/d_a_obj_wc_obj02.cpp#_ZN15daObjWc_Obj02_cD1Ev",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c10SetWaterIDEv",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c16CleanupResourcesEv",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c19CheckClsnWithPlayerEv",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c20UpdateModelTransformEv",
+    "src/actors/d_a_obj_wc_obj03.cpp#_ZN15daObjWc_Obj03_c6RenderEv",
     "src/actors/d_a_obj_wc_obj04.cpp#_ZN15daObjWc_Obj04_c16CleanupResourcesEv",
     "src/actors/d_a_obj_wc_obj04.cpp#_ZN15daObjWc_Obj04_c6RenderEv",
     "src/actors/d_a_obj_wc_obj05.cpp#_ZN15daObjWc_Obj05_c6RenderEv",
