--- src/_ZN15daObjWc_Obj03_c19CheckClsnWithPlayerEv.cpp
@@ -1,31 +1,0 @@
-//cpp
-#include "daObjWc_Obj03_c.h"
-#include "daObjWc_Mizu_c.h"
-
-struct Sound {
-    static void PlayBank3(u32 id, const Vector3 &pos);
-};
-
-void daObjWc_Obj03_c::CheckClsnWithPlayer()
-{
-    if (mActive != 0)
-        return;
-    u32 id = mCylinder.otherOwner;
-    if (id == 0)
-        return;
-    dActor_c *actor = dActor_c::FindWithID(id);
-    if (actor == 0)
-        return;
-    int isPlayer = (int)(actor->actorID == 0xbf);
-    if (isPlayer == 0)
-        return;
-    daObjWc_Mizu_c *water = (daObjWc_Mizu_c *)dActor_c::FindWithID(mWaterID);
-    int targetPosY = water->mTargetPosY;
-    if (targetPosY != water->mPosY)
-        return;
-    if (targetPosY == mPosY)
-        return;
-    mActive = 1;
-    water->mTargetPosY = mPosY;
-    Sound::PlayBank3(0x63, *(Vector3 *)&mCamSpacePosX);
-}
