--- src/_ZN15daObjWc_Obj03_cD0Ev.cpp
@@ -1,7 +1,0 @@
-//cpp
-// @symbol _ZN15daObjWc_Obj03_cD0Ev
-#include "daObjWc_Obj03_c.h"
-
-daObjWc_Obj03_c::~daObjWc_Obj03_c()
-{
-}
